Neopixel driver now supports 16 bit addressing.  Made stream consumer task more resilient to malformed events

+++ display/datalamp/program.h
@@ -11,7 +11,7 @@ struct program_config {
 };
 
 struct program_event {
-  uint8_t address;
+  uint16_t address;
   char data[PROGRAM_EVENT_DATA_SIZE];
 };
